refactor: unify contract version call
