+++ Cargo.toml
@@ -114,7 +114,6 @@ alloy = { version = "1.0.23", features = [
     "provider-anvil-node",
     "reqwest-rustls-tls",
 ], default-features = false }
-alloy-primitives = { version = "1.3.0", features = ["rand"] }
 derive_more = { version = "2", features = ["debug"] }
 dotenvy = "0.15"
 itertools = "0.14"
